Add armadillo linear algebra library
Add the header files for the c++ armadillo library. This will be
used to produce faster code, to ensure minimal runtimes. This
library takes advantage of BLAS and LAPACK, which must be insta-
lled on the system for it to work properly. Inclusion of these
files assumes that the armadillo library has been built and is
linked to when building.