Add Nearest Neighbor classifier constructor
Defines two constructors for the Nearest Neighbor class to provide
the necessary information to compute the distance metric using
armadillo. These constructors are an unsigned integer for the
p-norm metric and a string constructor for the use of the infinity
norms. The default value for the unsigned integer constructor is
2 for the euclidean norm. The string constructor will only work
for values "inf" for the infinity norm (max) and "-inf" for the
-infinity norm (min).

Corresponding tests are also defined. These ensure that the only
valid parameters are integer numerical types and the strings "inf"
and "-inf". Other string arguments should be invalid and produce
and error of type std::invalid_argument.