Fix predict method of nearest neighbor
The predict method in the nearest neighbor classifier had an error
in the line that computed the distance. Fixed the distance measure
to compute the distance between the desired vector and each column
of the training set rather than the entire training set.

Also removed the serialization routines which should have been
removed before.