Implement classifier benchmarking routine
Create a general benchmark routine that applies implemented
algorithms to the MNIST digit database for the purpose of
comparing the training and testing times, as well as the error
rates for the various algorithms.

The following algorithms were implemented:

- Nearest-Neighbor (Manhattan distance)
- Nearest-Neighbor (Euclidean distance)

A timer class was also created to assist in benchmarking the
routines and to reduce code length in the benchmarking routine.