Add PCA to benchmarking routine
Perform PCA on the MNIST dataset to reduce computational time and
increase accuracy. The number of dimensions specified, 56, was
determined previously in another project and is used here without
proof. If you wish to understand why this value was chosen,
perform calculations iterating through all possible number of
dimensions and it can be seen that 56 provides the greatest
accuracy.